+++ app/src/androidTest/java/com/forrestguice/suntimeswidget/settings/WidgetSettingsTest.java
@@ -420,6 +420,22 @@ public class WidgetSettingsTest extends SuntimesActivityTestBase
     }
 
     @Test
+    public void test_1x1SunPosModePref()
+    {
+        WidgetSettings.saveSunPos1x1ModePref(context, appWidgetId, WidgetSettings.WidgetModeSunPos1x1.MODE1x1_ALTAZ);
+        WidgetSettings.WidgetModeSunPos1x1 pref2 = WidgetSettings.loadSunPos1x1ModePref(context, appWidgetId);
+        assertTrue("pref should be ALTAZ but was " + pref2, pref2.equals(WidgetSettings.WidgetModeSunPos1x1.MODE1x1_ALTAZ));
+
+        WidgetSettings.saveSunPos1x1ModePref(context, appWidgetId, WidgetSettings.WidgetModeSunPos1x1.MODE1x1_DECRIGHT);
+        WidgetSettings.WidgetModeSunPos1x1 pref1 = WidgetSettings.loadSunPos1x1ModePref(context, appWidgetId);
+        assertTrue("pref should be DECRIGHT but was " + pref1, pref1.equals(WidgetSettings.WidgetModeSunPos1x1.MODE1x1_DECRIGHT));
+
+        WidgetSettings.deleteSunPos1x1ModePref(context, appWidgetId);
+        WidgetSettings.WidgetModeSunPos1x1 pref0 = WidgetSettings.loadSunPos1x1ModePref(context, appWidgetId);
+        assertTrue("pref should be default (ALTAZ) but was " + pref0, pref0.equals(WidgetSettings.PREF_DEF_APPEARANCE_WIDGETMODE_SUNPOS1x1) && pref0.equals(WidgetSettings.WidgetModeSunPos1x1.MODE1x1_ALTAZ));
+    }
+
+    @Test
     public void test_1x1MoonModePref()
     {
         WidgetSettings.saveMoon1x1ModePref(context, appWidgetId, WidgetSettings.WidgetModeMoon1x1.MODE1x1_RISESET);
@@ -549,6 +565,22 @@ public class WidgetSettingsTest extends SuntimesActivityTestBase
     }
 
     @Test
+    public void test_observerHeightPref()
+    {
+        WidgetSettings.saveObserverHeightPref(context, appWidgetId, 1f);
+        float height0 = WidgetSettings.loadObserverHeightPref(context, appWidgetId);
+        assertTrue("height should be 1 but was " + height0, equals(height0, 1f));
+
+        WidgetSettings.saveObserverHeightPref(context, appWidgetId, 2.5f);
+        float height1 = WidgetSettings.loadObserverHeightPref(context, appWidgetId);
+        assertTrue("height should be 2.5 but was " + height1, equals(height1, 2.5f));
+
+        WidgetSettings.deleteObserverHeightPref(context, appWidgetId);
+        float height2 = WidgetSettings.loadObserverHeightPref(context, appWidgetId);
+        assertTrue("height should be default (1.8288) but was " + height2, equals(height2, 1.8288f) && equals(height2, WidgetSettings.PREF_DEF_GENERAL_OBSERVERHEIGHT));
+    }
+
+    @Test
     public void test_showTimeDatePref()
     {
         WidgetSettings.saveShowTimeDatePref(context, appWidgetId, true);
@@ -564,5 +596,10 @@ public class WidgetSettingsTest extends SuntimesActivityTestBase
         assertTrue("showTimeDate should be default (true) but was " + showTimeDate, showTimeDate && showTimeDate == WidgetSettings.PREF_DEF_GENERAL_SHOWTIMEDATE);
     }
 
+    public static final float FLOAT_TOLERANCE = 0.01f;
+    protected boolean equals(float float1, float float2)
+    {
+        return (Math.abs(float1 - float2) < FLOAT_TOLERANCE);
+    }
 
 }