+++ app/src/main/java/com/forrestguice/suntimeswidget/calculator/ca/rmen/sunrisesunset/SunriseSunsetSuntimesCalculator.java
@@ -243,6 +243,12 @@ public class SunriseSunsetSuntimesCalculator implements SuntimesCalculator
     }
 
     @Override
+    public SunPosition getSunPosition(Calendar dateTime)
+    {
+        return null;
+    }
+
+    @Override
     public Calendar getOfficialSunsetCalendarForDate( Calendar date )
     {
         Calendar[] riseset = SunriseSunset.getSunriseSunset(date, location.getLatitudeAsDouble(), location.getLongitudeAsDouble());